+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - L2 AmeriFlux formatting variables to env editor. [#126](https://github.com/ncsa/ameriflux-pipeline/issues/126)
 - Formatting of PyFluxPro output for Ameriflux. [#130](https://github.com/ncsa/ameriflux-pipeline/issues/130)
 - Reading precip data more robust. [#137](https://github.com/ncsa/ameriflux-pipeline/issues/137)
+- Merging to various raw met data files to csv format for a specified time period. [#134](https://github.com/ncsa/ameriflux-pipeline/issues/134)
 
 ### Fixed
 - Python packaging of the codes so it can run in command prompt. [#70](https://github.com/ncsa/ameriflux-pipeline/issues/70)